requirements.txt
streamlit==1.51.0
pandas==2.3.3
altair==5.5.0
openpyxl==3.1.5
sqlalchemy==2.0.44
schedule==1.2.2
gitpython==3.1.45
